Commit changes for walkthrough 09

+++ src/calculator.cpp
@@ -1,5 +1,6 @@
 #include "include/calculator.h"
 #include <cmath>
+#include <iostream>
 int Calculator::Add(int a, int b) 
 {
     return a+b;
@@ -14,7 +15,7 @@ double Calculator::CalculateInterest(const double principal, const std::string& 
 double Calculator::CalculateInterest(const double principal, const std::string& type, const int term, const int compounded) const
 {
     int times = term / compounded;
-    int years = term / 365;
+    double years = term / 365.0;
     double rate = rates.at(type);
 
     double amount = principal * std::pow((1 + rate/times),(times*years));
